+++ pylint/checkers/format.py
@@ -52,8 +52,8 @@ from astroid import nodes
 
 from pylint.checkers import BaseTokenChecker
 from pylint.checkers.utils import check_messages
+from pylint.constants import OPTION_RGX, WarningScope
 from pylint.interfaces import IAstroidChecker, IRawChecker, ITokenChecker
-from pylint.utils import OPTION_RGX, WarningScope
 
 _ASYNC_TOKEN = "async"
 _CONTINUATION_BLOCK_OPENERS = [
