+++ pylint/constants.py
@@ -3,6 +3,15 @@
 # Licensed under the GPL: https://www.gnu.org/licenses/old-licenses/gpl-2.0.html
 # For details: https://github.com/PyCQA/pylint/blob/master/COPYING
 
+import re
+
+# Allow stopping after the first semicolon/hash encountered,
+# so that an option can be continued with the reasons
+# why it is active or disabled.
+OPTION_RGX = re.compile(r"\s*#.*\bpylint:\s*([^;#]+)[;#]{0,1}")
+
+PY_EXTS = (".py", ".pyc", ".pyo", ".pyw", ".so", ".dll")
+
 MSG_STATE_CONFIDENCE = 2
 _MSG_ORDER = "EWRCIF"
 MSG_STATE_SCOPE_CONFIG = 0
@@ -22,3 +31,8 @@ MSG_TYPES = {
 MSG_TYPES_LONG = {v: k for k, v in MSG_TYPES.items()}
 
 MSG_TYPES_STATUS = {"I": 0, "C": 16, "R": 8, "W": 4, "E": 2, "F": 1}
+
+
+class WarningScope:
+    LINE = "line-based-msg"
+    NODE = "node-based-msg"
