+++ pylint/message/build_message_definition.py
@@ -5,9 +5,9 @@
 
 import warnings
 
+from pylint.constants import WarningScope
 from pylint.interfaces import IRawChecker, ITokenChecker, implements
 from pylint.message.message_definition import MessageDefinition
-from pylint.utils.warning_scope import WarningScope
 
 
 def build_message_def(checker, msgid, msg_tuple):
