+++ pylint/utils/file_state.py
@@ -7,8 +7,7 @@ import collections
 
 from astroid import nodes
 
-from pylint.message.constants import MSG_STATE_SCOPE_MODULE
-from pylint.utils.warning_scope import WarningScope
+from pylint.constants import MSG_STATE_SCOPE_MODULE, WarningScope
 
 
 class FileState:
