--- pylint/utils/warning_scope.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Licensed under the GPL: https://www.gnu.org/licenses/old-licenses/gpl-2.0.html
-# For details: https://github.com/PyCQA/pylint/blob/master/COPYING
-
-
-class WarningScope:
-    LINE = "line-based-msg"
-    NODE = "node-based-msg"
